+++ src/dataset.rs
@@ -13,7 +13,7 @@ pub type Item = Vec<Attr>;
 
 #[derive(Debug)]
 pub struct Attr {
-    pub id: u32,
+    pub id: usize,
     pub value: f64,
 }
 
@@ -75,7 +75,7 @@ impl From<Lines<'_>> for Dataset {
                     let item: Item = attrs
                         .split('\t')
                         .map(|s| Attr {
-                            id: m_attrs.find_or_insert(s) as u32,
+                            id: m_attrs.find_or_insert(s),
                             value: 1.0,
                         })
                         .collect();
@@ -109,7 +109,7 @@ impl TryFrom<File> for Dataset {
                     let item: Item = attrs
                         .split('\t')
                         .map(|s| Attr {
-                            id: m_attrs.find_or_insert(s) as u32,
+                            id: m_attrs.find_or_insert(s),
                             value: 1.0,
                         })
                         .collect();
@@ -138,7 +138,7 @@ impl Dataset {
                     let item: Item = attrs
                         .split('\t')
                         .map(|s| Attr {
-                            id: self.m_attrs.find_or_insert(s) as u32,
+                            id: self.m_attrs.find_or_insert(s),
                             value: 1.0,
                         })
                         .collect();
